ASd33

+++ resources/views/index/statement.blade.php
@@ -53,7 +53,7 @@
                 <div class="col-md-2 col-md-offset-2 text-center xs-margin-bottom-ten col-sm-4">
                     <div class="special-gifts-box ">
                         <span class="text-uppercase text-small letter-spacing-1">Total Donation</span><br/>
-                        <span class="gifts-off bg-fast-blue white-text">৳ {{ (int) ($totaldonationamount->total - $totalcharge->total) }}</span>
+                        <span class="gifts-off bg-fast-blue white-text">৳ {{ $totaldonationamount->total - $totalcharge->total }}</span>
                     </div>
                 </div>
                 <div class="col-md-2 text-center xs-margin-bottom-ten col-sm-4">
